Move cards code into seperate module
